(lib) move transition type internally
This reduces the C-ABI-specific parts of the surface API down to a handful
of step functions.

Morally, there is still only one step function inside with a transition datatype,
but the API exposes them via multiple functions.
