+++ src/lib/casemate.h
@@ -19,32 +19,6 @@
 #define CASEMATE_VERSION "2.0.2"
 
 /* auto-included by Makefile */
-/* Types also defined by UoC's pKVM ghost code headers
- * so do not try to include them in the top-level casemate.h if they are already defined
- */
-
-#ifndef __KVM_NVHE_HYPERVISOR__
-
-#include <stdarg.h>
-#include <stdint.h>
-#include <stdbool.h>
-
-typedef unsigned long u64;
-typedef signed long s64;
-typedef unsigned int u32;
-typedef signed int s32;
-typedef int u8;
-typedef u64 phys_addr_t;
-
-#else
-
-#include <linux/stdarg.h>
-#include <linux/types.h>
-
-#endif /* __KVM_NVHE_HYPERVISOR */
-
-
-/* auto-included by Makefile */
 #ifndef CASEMATE_CONFIG_H
 #define CASEMATE_CONFIG_H
 
@@ -175,8 +149,8 @@ enum ghost_sysreg_kind {
 };
 
 struct casemate_model_step;
-typedef int (*vprintf_cb)(void* arg, const char *format, va_list ap);
-typedef void* (*sprint_make_buf_cb)(char* arg, uint64_t n);
+typedef int (*vprintf_cb)(void *arg, const char *format, va_list ap);
+typedef void *(*sprint_make_buf_cb)(char *arg, uint64_t n);
 typedef void (*sprint_free_buf_cb)(void *buf);
 typedef void (*abort_cb)(const char *msg);
 typedef uint64_t (*read_physmem_cb)(uint64_t);
@@ -265,45 +239,6 @@ enum sm_tlbi_op_regime_kind {
 	TLBI_REGIME_EL2 = 2, /* EL2 regime */
 };
 
-/**
- * struct sm_tlbi_op_method - Decoded TLBI by-method
- * @kind: whether this is "all", by address, or by address-space-identifier.
- */
-struct sm_tlbi_op_method {
-	enum sm_tlbi_op_method_kind kind;
-	union {
-		struct tlbi_op_method_by_address_data {
-			uint64_t page;
-
-			bool has_level_hint;
-			uint8_t level_hint;
-
-			bool has_asid;
-			uint8_t asid;
-
-			bool affects_last_level_only;
-		} by_address_data;
-
-		struct tlbi_op_method_by_address_space_id_data {
-			uint64_t asid_or_vmid;
-		} by_id_data;
-	};
-};
-
-/**
- * struct sm_tlbi_op - A decoded TLB maintenance operation.
- * @stage: whether this affects cached stage1 translations, cached stage2 translations, or both.
- * @regime: the translation regime that this TLB maintenance operation would affect the cached entries of.
- * @method: whether this TLBI is by IPA, by VA, or by VMID, etc., and the relevant address/vmid, etc.
- * @shootdown: whether this TLB maintenance operation is broadcast to other cores.
- */
-struct sm_tlbi_op {
-	enum sm_tlbi_op_stage stage;
-	enum sm_tlbi_op_regime_kind regime;
-	struct sm_tlbi_op_method method;
-	bool shootdown;
-};
-
 /// Encoded TLBIs
 
 enum tlbi_kind {
@@ -328,110 +263,6 @@ enum barrier_kind {
 	BARRIER_ISB
 };
 
-enum casemate_model_step_kind {
-	/**
-	 * @TRANS_HW_STEP - Hardware instruction
-	 */
-	TRANS_HW_STEP,
-
-	/**
-	 * @TRANS_ABS_STEP - An abstract software state transition
-	 * These generally transition some abstract reified ghost state in the model
-	 * e.g. for locks and C initialisations that are not exposed to the model
-	 */
-	TRANS_ABS_STEP,
-
-	/**
-	 * @TRANS_HINT - A non-hardware-model transition
-	 * These generally provide additional information to the casemate model,
-	 * such as ownership, to resolve otherwise unbounded non-determinism
-	 *
-	 * Removing HINTs should not change soundness
-	 */
-	TRANS_HINT,
-};
-
-enum ghost_hw_step_kind {
-	HW_MEM_WRITE,
-	HW_MEM_READ,
-	HW_BARRIER,
-	HW_TLBI,
-	HW_MSR,
-};
-
-struct ghost_hw_step {
-	enum ghost_hw_step_kind kind;
-	union {
-		struct trans_write_data {
-			enum memory_order_t mo;
-			uint64_t phys_addr;
-			uint64_t val;
-		} write_data;
-
-		struct trans_read_data {
-			uint64_t phys_addr;
-			uint64_t val;
-		} read_data;
-
-		struct trans_barrier_data {
-			enum barrier_kind kind;
-			enum dxb_kind dxb_data;
-		} barrier_data;
-
-		struct trans_tlbi_data {
-			enum tlbi_kind tlbi_kind;
-			uint64_t value;
-		} tlbi_data;
-
-		struct trans_msr_data {
-			enum ghost_sysreg_kind sysreg;
-			uint64_t val;
-		} msr_data;
-	};
-};
-
-enum ghost_abs_kind {
-	/**
-	 * @GHOST_ABS_LOCK - Acquire a mutex
-	 */
-	GHOST_ABS_LOCK,
-
-	/**
-	 * @GHOST_ABS_UNLOCK - Release a mutex
-	 */
-	GHOST_ABS_UNLOCK,
-
-	/**
-	 * @GHOST_ABS_INIT - Zeroed initialisation of some fresh memory
-	 */
-	GHOST_ABS_INIT,
-
-	/**
-	 * @GHOST_ABS_MEMSET - A C memset() call.
-	 */
-	GHOST_ABS_MEMSET,
-};
-
-struct ghost_abs_step {
-	enum ghost_abs_kind kind;
-	union {
-		struct trans_init_data {
-			uint64_t location;
-			uint64_t size;
-		} init_data;
-
-		struct trans_lock_data {
-			uint64_t address;
-		} lock_data;
-
-		struct trans_memset_data {
-			uint64_t address;
-			uint64_t size;
-			uint64_t value;
-		} memset_data;
-	};
-};
-
 enum ghost_hint_kind {
 	/**
 	 * @GHOST_HINT_SET_ROOT_LOCK - Set the lock owning a pgtable root.
@@ -454,12 +285,6 @@ enum ghost_hint_kind {
 	GHOST_HINT_SET_PTE_THREAD_OWNER,
 };
 
-struct ghost_hint_step {
-	enum ghost_hint_kind kind;
-	uint64_t location;
-	uint64_t value;
-};
-
 /**
  * Source location info
  */
@@ -467,32 +292,6 @@ struct src_loc {
 	const char *file;
 	const char *func;
 	int lineno;
-};
-
-struct casemate_model_step {
-	/**
-	 * @tid: thread identifier.
-	 */
-	uint8_t tid;
-
-	/**
-	 * @seq_id: sequence id number of the transition.
-	 */
-	uint64_t seq_id;
-
-	/**
-	 * @src_loc: string location (path, function name, lineno etc)
-	 *           of where the transition happens in the source code.
-	 *           For debugging/pretty printing.
-	 */
-	struct src_loc src_loc;
-
-	enum casemate_model_step_kind kind;
-	union {
-		struct ghost_hw_step hw_step;
-		struct ghost_abs_step abs_step;
-		struct ghost_hint_step hint_step;
-	};
 };
 
 /**
@@ -509,16 +308,18 @@ struct casemate_model_step {
  *
  * NOTE: After this the target must manually initialise the already-existing pagetable memory with steps.
  */
-int initialise_casemate_model(struct casemate_options *opts, uint64_t phys, uint64_t size, void *sm_virt, uint64_t sm_size);
+int initialise_casemate_model(struct casemate_options *opts, uint64_t phys, uint64_t size,
+			      void *sm_virt, uint64_t sm_size);
 
 /**
- * casemate_model_step() - Take a step in the ghost model.
+ * casemate_cpu_id() - Return current CPU identifier.
+ *
+ * Users should implement this if they want to use the helper macros.
  */
-void casemate_model_step(struct casemate_model_step trans);
+extern uint64_t casemate_cpu_id(void);
 
+/// Transition API Steps ///
 
-/* auto-included by Makefile */
-//////////////
 // Step helpers
 
 #define SRC_LOC \
@@ -527,242 +328,61 @@ void casemate_model_step(struct casemate_model_step trans);
 		.file = __FILE__, .lineno = __LINE__, .func = __func__ \
 	}
 
-/**
- * casemate_cpu_id() - Return current CPU identifier.
- *
- * Users should implement this if they want to use the helper macros.
- */
-extern uint64_t casemate_cpu_id(void);
 #define THREAD_ID casemate_cpu_id()
 
-#define casemate_model_step_write(...) __casemate_model_step_write(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_write(uint64_t tid, struct src_loc src_loc, enum memory_order_t mo, uint64_t phys, uint64_t val)
-{
-	casemate_model_step((struct casemate_model_step){
-		.tid = tid,
-		.src_loc = src_loc,
-		.kind = TRANS_HW_STEP,
-		.hw_step =
-			(struct ghost_hw_step){
-				.kind = HW_MEM_WRITE,
-				.write_data =
-					(struct trans_write_data){
-						.mo = mo,
-						.phys_addr = phys,
-						.val = val,
-					},
-			},
-	});
-}
+#define casemate_model_step_write(...) \
+	__casemate_model_step_write(THREAD_ID, SRC_LOC, __VA_ARGS__)
+void __casemate_model_step_write(uint64_t tid, struct src_loc src_loc, enum memory_order_t mo,
+				 uint64_t phys, uint64_t val);
 
 #define casemate_model_step_read(...) __casemate_model_step_read(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_read(uint64_t tid, struct src_loc src_loc, uint64_t phys, uint64_t val)
-{
-	casemate_model_step((struct casemate_model_step){
-		.tid = tid,
-		.src_loc = src_loc,
-		.kind = TRANS_HW_STEP,
-		.hw_step =
-			(struct ghost_hw_step){
-				.kind = HW_MEM_READ,
-				.read_data =
-					(struct trans_read_data){
-						.phys_addr = phys,
-						.val = val,
-					},
-			},
-	});
-}
+void __casemate_model_step_read(uint64_t tid, struct src_loc src_loc, uint64_t phys,
+				uint64_t val);
 
 #define casemate_model_step_dsb(...) __casemate_model_step_dsb(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_dsb(uint64_t tid, struct src_loc src_loc, enum dxb_kind kind)
-{
-	casemate_model_step((struct casemate_model_step){
-		.tid = tid,
-		.src_loc = src_loc,
-		.kind = TRANS_HW_STEP,
-		.hw_step =
-			(struct ghost_hw_step){
-				.kind = HW_BARRIER,
-				.barrier_data =
-					(struct trans_barrier_data){
-						.kind = BARRIER_DSB,
-						.dxb_data = kind,
-					},
-			},
-	});
-}
+void __casemate_model_step_dsb(uint64_t tid, struct src_loc src_loc, enum dxb_kind kind);
 
 #define casemate_model_step_isb() __casemate_model_step_isb(THREAD_ID, SRC_LOC)
-static inline void __casemate_model_step_isb(uint64_t tid, struct src_loc src_loc)
-{
-	casemate_model_step((struct casemate_model_step){
-		.tid = tid,
-		.src_loc = src_loc,
-		.kind = TRANS_HW_STEP,
-		.hw_step =
-			(struct ghost_hw_step){
-				.kind = HW_BARRIER,
-				.barrier_data =
-					(struct trans_barrier_data){
-						.kind = BARRIER_ISB,
-					},
-			},
-	});
-}
+void __casemate_model_step_isb(uint64_t tid, struct src_loc src_loc);
 
-#define casemate_model_step_tlbi_reg(...) __casemate_model_step_tlbi_reg(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_tlbi_reg(uint64_t tid, struct src_loc src_loc, enum tlbi_kind kind, uint64_t value)
-{
-	casemate_model_step((struct casemate_model_step){
-		.tid = tid,
-		.src_loc = src_loc,
-		.kind = TRANS_HW_STEP,
-		.hw_step =
-			(struct ghost_hw_step){
-				.kind = HW_TLBI,
-				.tlbi_data =
-					(struct trans_tlbi_data){
-						.tlbi_kind = kind,
-						.value = value,
-					},
-			},
-	});
-}
+#define casemate_model_step_tlbi_reg(...) \
+	__casemate_model_step_tlbi_reg(THREAD_ID, SRC_LOC, __VA_ARGS__)
+
+void __casemate_model_step_tlbi_reg(uint64_t tid, struct src_loc src_loc, enum tlbi_kind kind,
+				    uint64_t value);
 
 #define casemate_model_step_tlbi(...) __casemate_model_step_tlbi(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_tlbi(uint64_t tid, struct src_loc src_loc, enum tlbi_kind kind)
-{
-	casemate_model_step((struct casemate_model_step){
-		.tid = tid,
-		.src_loc = src_loc,
-		.kind = TRANS_HW_STEP,
-		.hw_step =
-			(struct ghost_hw_step){
-				.kind = HW_TLBI,
-				.tlbi_data =
-					(struct trans_tlbi_data){
-						.tlbi_kind = kind,
-					},
-			},
-	});
-}
-
 #define casemate_model_step_tlbi_va(TLBI_KIND, ADDR, TTL, ASID) \
 	casemate_model_step_tlbi_reg((TLBI_KIND), (ADDR) | ((TTL) << 44ULL) | ((ASID) << 48ULL))
 
 #define casemate_model_step_tlbi_ipa(TLBI_KIND, ADDR, TTL) \
 	casemate_model_step_tlbi_reg((TLBI_KIND), (ADDR) | ((TTL) << 44ULL))
 
+void __casemate_model_step_tlbi(uint64_t tid, struct src_loc src_loc, enum tlbi_kind kind);
+
 #define casemate_model_step_msr(...) __casemate_model_step_msr(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_msr(uint64_t tid, struct src_loc src_loc, enum ghost_sysreg_kind sysreg, uint64_t val)
-{
-	casemate_model_step((struct casemate_model_step){
-		.tid = tid,
-		.src_loc = src_loc,
-		.kind = TRANS_HW_STEP,
-		.hw_step =
-			(struct ghost_hw_step){
-				.kind = HW_MSR,
-				.msr_data =
-					(struct trans_msr_data){
-						.sysreg = sysreg,
-						.val = val,
-					},
-			},
-	});
-}
+void __casemate_model_step_msr(uint64_t tid, struct src_loc src_loc,
+			       enum ghost_sysreg_kind sysreg, uint64_t val);
 
 #define casemate_model_step_hint(...) __casemate_model_step_hint(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_hint(uint64_t tid, struct src_loc src_loc, enum ghost_hint_kind kind, uint64_t location, uint64_t value)
-{
-	casemate_model_step((struct casemate_model_step){
-		.tid = tid,
-		.src_loc = src_loc,
-		.kind = TRANS_HINT,
-		.hint_step =
-			(struct ghost_hint_step){
-				.kind = kind,
-				.location = location,
-				.value = value,
-			},
-	});
-}
+void __casemate_model_step_hint(uint64_t tid, struct src_loc src_loc, enum ghost_hint_kind kind,
+				uint64_t location, uint64_t value);
 
 #define casemate_model_step_init(...) __casemate_model_step_init(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_init(uint64_t tid, struct src_loc src_loc, uint64_t location, uint64_t size)
-{
-	casemate_model_step((struct casemate_model_step){
-		.tid = tid,
-		.src_loc = src_loc,
-		.kind = TRANS_ABS_STEP,
-		.abs_step =
-			(struct ghost_abs_step){
-				.kind = GHOST_ABS_INIT,
-				.init_data =
-					(struct trans_init_data){
-						.location = location,
-						.size = size,
-					},
-			},
-	});
-}
+void __casemate_model_step_init(uint64_t tid, struct src_loc src_loc, uint64_t location,
+				uint64_t size);
 
-#define casemate_model_step_memset(...) __casemate_model_step_memset(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_memset(uint64_t tid, struct src_loc src_loc, uint64_t location, uint64_t value, uint64_t size)
-{
-	casemate_model_step((struct casemate_model_step){
-		.tid = tid,
-		.src_loc = src_loc,
-		.kind = TRANS_ABS_STEP,
-		.abs_step =
-			(struct ghost_abs_step){
-				.kind = GHOST_ABS_MEMSET,
-				.memset_data =
-					(struct trans_memset_data){
-						.address = location,
-						.size = size,
-						.value = value,
-					},
-			},
-	});
-}
+#define casemate_model_step_memset(...) \
+	__casemate_model_step_memset(THREAD_ID, SRC_LOC, __VA_ARGS__)
+void __casemate_model_step_memset(uint64_t tid, struct src_loc src_loc, uint64_t location,
+				  uint64_t value, uint64_t size);
 
 #define casemate_model_step_lock(...) __casemate_model_step_lock(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_lock(uint64_t tid, struct src_loc src_loc, uint64_t address)
-{
-	casemate_model_step((struct casemate_model_step){
-		.tid = tid,
-		.src_loc = src_loc,
-		.kind = TRANS_ABS_STEP,
-		.abs_step =
-			(struct ghost_abs_step){
-				.kind = GHOST_ABS_LOCK,
-				.lock_data =
-					(struct trans_lock_data){
-						.address = address,
-					},
-			},
-	});
-}
+void __casemate_model_step_lock(uint64_t tid, struct src_loc src_loc, uint64_t address);
 
-#define casemate_model_step_unlock(...) __casemate_model_step_unlock(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_unlock(uint64_t tid, struct src_loc src_loc, uint64_t address)
-{
-	casemate_model_step((struct casemate_model_step){
-		.tid = tid,
-		.src_loc = src_loc,
-		.kind = TRANS_ABS_STEP,
-		.abs_step =
-			(struct ghost_abs_step){
-				.kind = GHOST_ABS_UNLOCK,
-				.lock_data =
-					(struct trans_lock_data){
-						.address = address,
-					},
-			},
-	});
-}
+#define casemate_model_step_unlock(...) \
+	__casemate_model_step_unlock(THREAD_ID, SRC_LOC, __VA_ARGS__)
+void __casemate_model_step_unlock(uint64_t tid, struct src_loc src_loc, uint64_t address);
 
 
 #endif /* CASEMATE_H */
