+++ src/lib/include/casemate-impl.h
@@ -23,6 +23,7 @@
 #include <casemate-impl/arch.h>
 #include <casemate-impl/pgtable.h>
 
+#include <casemate-impl/transitions.h>
 #include <casemate-impl/state.h>
 
 #include <casemate-impl/model.h>
