+++ src/lib/include/casemate-in/casemate.out.h
@@ -22,7 +22,4 @@ CASEMATE_CONFIG_H
 /* auto-included by Makefile */
 CASEMATE_TRANSITIONS_H
 
-/* auto-included by Makefile */
-CASEMATE_STEP_HELPERS_H
-
 #endif /* CASEMATE_H */
