+++ src/lib/src/transitions.c
@@ -1,25 +1,9 @@
-//////////////
-// Step helpers
+#include <casemate.h>
 
-#define SRC_LOC \
-	(struct src_loc) \
-	{ \
-		.file = __FILE__, .lineno = __LINE__, .func = __func__ \
-	}
+#include <casemate-impl.h>
 
-/**
- * casemate_cpu_id() - Return current CPU identifier.
- *
- * Users should implement this if they want to use the helper macros.
- */
-extern uint64_t casemate_cpu_id(void);
-#define THREAD_ID casemate_cpu_id()
-
-#define casemate_model_step_write(...) \
-	__casemate_model_step_write(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_write(uint64_t tid, struct src_loc src_loc,
-					       enum memory_order_t mo, uint64_t phys,
-					       uint64_t val)
+void __casemate_model_step_write(uint64_t tid, struct src_loc src_loc, enum memory_order_t mo,
+				 uint64_t phys, uint64_t val)
 {
 	casemate_model_step((struct casemate_model_step){
 		.tid = tid,
@@ -38,9 +22,7 @@ static inline void __casemate_model_step_write(uint64_t tid, struct src_loc src_
 	});
 }
 
-#define casemate_model_step_read(...) __casemate_model_step_read(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_read(uint64_t tid, struct src_loc src_loc, uint64_t phys,
-					      uint64_t val)
+void __casemate_model_step_read(uint64_t tid, struct src_loc src_loc, uint64_t phys, uint64_t val)
 {
 	casemate_model_step((struct casemate_model_step){
 		.tid = tid,
@@ -58,9 +40,7 @@ static inline void __casemate_model_step_read(uint64_t tid, struct src_loc src_l
 	});
 }
 
-#define casemate_model_step_dsb(...) __casemate_model_step_dsb(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_dsb(uint64_t tid, struct src_loc src_loc,
-					     enum dxb_kind kind)
+void __casemate_model_step_dsb(uint64_t tid, struct src_loc src_loc, enum dxb_kind kind)
 {
 	casemate_model_step((struct casemate_model_step){
 		.tid = tid,
@@ -78,8 +58,7 @@ static inline void __casemate_model_step_dsb(uint64_t tid, struct src_loc src_lo
 	});
 }
 
-#define casemate_model_step_isb() __casemate_model_step_isb(THREAD_ID, SRC_LOC)
-static inline void __casemate_model_step_isb(uint64_t tid, struct src_loc src_loc)
+void __casemate_model_step_isb(uint64_t tid, struct src_loc src_loc)
 {
 	casemate_model_step((struct casemate_model_step){
 		.tid = tid,
@@ -96,10 +75,8 @@ static inline void __casemate_model_step_isb(uint64_t tid, struct src_loc src_lo
 	});
 }
 
-#define casemate_model_step_tlbi_reg(...) \
-	__casemate_model_step_tlbi_reg(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_tlbi_reg(uint64_t tid, struct src_loc src_loc,
-						  enum tlbi_kind kind, uint64_t value)
+void __casemate_model_step_tlbi_reg(uint64_t tid, struct src_loc src_loc, enum tlbi_kind kind,
+				    uint64_t value)
 {
 	casemate_model_step((struct casemate_model_step){
 		.tid = tid,
@@ -117,9 +94,7 @@ static inline void __casemate_model_step_tlbi_reg(uint64_t tid, struct src_loc s
 	});
 }
 
-#define casemate_model_step_tlbi(...) __casemate_model_step_tlbi(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_tlbi(uint64_t tid, struct src_loc src_loc,
-					      enum tlbi_kind kind)
+void __casemate_model_step_tlbi(uint64_t tid, struct src_loc src_loc, enum tlbi_kind kind)
 {
 	casemate_model_step((struct casemate_model_step){
 		.tid = tid,
@@ -136,15 +111,8 @@ static inline void __casemate_model_step_tlbi(uint64_t tid, struct src_loc src_l
 	});
 }
 
-#define casemate_model_step_tlbi_va(TLBI_KIND, ADDR, TTL, ASID) \
-	casemate_model_step_tlbi_reg((TLBI_KIND), (ADDR) | ((TTL) << 44ULL) | ((ASID) << 48ULL))
-
-#define casemate_model_step_tlbi_ipa(TLBI_KIND, ADDR, TTL) \
-	casemate_model_step_tlbi_reg((TLBI_KIND), (ADDR) | ((TTL) << 44ULL))
-
-#define casemate_model_step_msr(...) __casemate_model_step_msr(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_msr(uint64_t tid, struct src_loc src_loc,
-					     enum ghost_sysreg_kind sysreg, uint64_t val)
+void __casemate_model_step_msr(uint64_t tid, struct src_loc src_loc,
+			       enum ghost_sysreg_kind sysreg, uint64_t val)
 {
 	casemate_model_step((struct casemate_model_step){
 		.tid = tid,
@@ -162,10 +130,8 @@ static inline void __casemate_model_step_msr(uint64_t tid, struct src_loc src_lo
 	});
 }
 
-#define casemate_model_step_hint(...) __casemate_model_step_hint(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_hint(uint64_t tid, struct src_loc src_loc,
-					      enum ghost_hint_kind kind, uint64_t location,
-					      uint64_t value)
+void __casemate_model_step_hint(uint64_t tid, struct src_loc src_loc, enum ghost_hint_kind kind,
+				uint64_t location, uint64_t value)
 {
 	casemate_model_step((struct casemate_model_step){
 		.tid = tid,
@@ -180,9 +146,8 @@ static inline void __casemate_model_step_hint(uint64_t tid, struct src_loc src_l
 	});
 }
 
-#define casemate_model_step_init(...) __casemate_model_step_init(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_init(uint64_t tid, struct src_loc src_loc,
-					      uint64_t location, uint64_t size)
+void __casemate_model_step_init(uint64_t tid, struct src_loc src_loc, uint64_t location,
+				uint64_t size)
 {
 	casemate_model_step((struct casemate_model_step){
 		.tid = tid,
@@ -200,10 +165,8 @@ static inline void __casemate_model_step_init(uint64_t tid, struct src_loc src_l
 	});
 }
 
-#define casemate_model_step_memset(...) \
-	__casemate_model_step_memset(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_memset(uint64_t tid, struct src_loc src_loc,
-						uint64_t location, uint64_t value, uint64_t size)
+void __casemate_model_step_memset(uint64_t tid, struct src_loc src_loc, uint64_t location,
+				  uint64_t value, uint64_t size)
 {
 	casemate_model_step((struct casemate_model_step){
 		.tid = tid,
@@ -222,9 +185,7 @@ static inline void __casemate_model_step_memset(uint64_t tid, struct src_loc src
 	});
 }
 
-#define casemate_model_step_lock(...) __casemate_model_step_lock(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_lock(uint64_t tid, struct src_loc src_loc,
-					      uint64_t address)
+void __casemate_model_step_lock(uint64_t tid, struct src_loc src_loc, uint64_t address)
 {
 	casemate_model_step((struct casemate_model_step){
 		.tid = tid,
@@ -241,10 +202,7 @@ static inline void __casemate_model_step_lock(uint64_t tid, struct src_loc src_l
 	});
 }
 
-#define casemate_model_step_unlock(...) \
-	__casemate_model_step_unlock(THREAD_ID, SRC_LOC, __VA_ARGS__)
-static inline void __casemate_model_step_unlock(uint64_t tid, struct src_loc src_loc,
-						uint64_t address)
+void __casemate_model_step_unlock(uint64_t tid, struct src_loc src_loc, uint64_t address)
 {
 	casemate_model_step((struct casemate_model_step){
 		.tid = tid,
